refactor: extract helpers/hardware.py (Phase 6.4)
Move `_detect_host_hardware()` out of dashboard.py. Pure cross-platform
detection function — sysctl on macOS, /proc on Linux, wmic on Windows.
Returns {cpu, cores, ram_gb, backend}. No module-level state.

Call sites: routes/infra.py:548 (cost-optimizer panel fallback),
routes/infra.py:689 (automation-analysis system block).

History: this function replaced a hardcoded "Apple M2 Pro / 12 / 32 GB"
fallback in PR #625 after the user caught it via screenshot
("is anything hardcoded here? like apple m2?"). The fallback had
happened to match reality on the user's MacBook (so nobody noticed
locally), but misrepresented every Linux/Windows install.

## E2E

- Unit test: live call on this Mac returns cpu=Apple M2 Pro (real
  sysctl output), cores=12, ram_gb=32, backend=Apple Metal (unified)
- /api/cost-optimizer system block: cpu=Apple M2 Pro, cores=12,
  ram_gb=32.0, backend=Apple Metal (unified)
- /api/component/machine: Hostname=Macbook-Pro.local, IP=192.168.178.56,
  CPU=arm, Cores=12
- 39/39 API endpoints return 200
- Zero tracebacks

## Sizes

  dashboard.py         : -80 lines (two identical 80-line duplicates
                                    removed; 1-line re-export added)
  helpers/hardware.py  : new (90 lines with docstrings)
